commit on the new git

+++ src/Controller/DepositController.php
@@ -4,6 +4,8 @@ namespace App\Controller;
 
 use App\Service\UploadFile;
 use App\Service\ZenodoClient;
+use Monolog\Logger;
+use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,7 +16,6 @@ use Symfony\Component\HttpClient\HttpClient;
 use GuzzleHttp\Client as guzzleClient;
 use App\Form\DepositFormType;
 use Symfony\Component\Security\Core\Security;
-use Symfony\Component\String\Slugger\SluggerInterface;
 use App\Repository\LogUserActionRepository;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
@@ -22,14 +23,11 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 class DepositController extends AbstractController
 {
 
-    public function new(Request $request, SluggerInterface $slugger, Security $security, LogUserActionRepository $logRepo, ZenodoClient $zenodoClient, UploadFile $uploadFile): Response
+    public function new(Request $request, Security $security, LogUserActionRepository $logRepo, ZenodoClient $zenodoClient, UploadFile $uploadFile,LoggerInterface $logger): Response
     {
-
         // token from CAS
         $userInfo = $security->getToken()->getAttributes();
-        //build form
         $form = $this->createForm(DepositFormType::class);
-        //Get request
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $deposit = $form->getData();
@@ -49,25 +47,40 @@ class DepositController extends AbstractController
                 if(!is_null($depositFile)) {
                     $zenodoClient->postFileInDeposit($depositFile,$tmpResponse,$this->getParameter('app.SBX_TOKEN'),$this->getParameter('deposit_upload_directory'));
                 }
-                $zenodoClient->postMetadataInDeposit($deposit,$idDeposit,$this->getParameter('app.SBX_TOKEN'));
-
-                $action = 'save';
-                if ($form->getClickedButton() && 'save_publish' === $form->getClickedButton()->getName()) {
-                    $zenodoClient->publishDeposit($idDeposit,$this->getParameter('app.SBX_TOKEN'));
-                    $action = 'publish';
+                $postMetadatas = $zenodoClient->postMetadataInDeposit($deposit,$idDeposit,$this->getParameter('app.SBX_TOKEN'));
+                if ($postMetadatas->getStatusCode() === 200) {
+                    $action = 'save';
+                    if ($form->getClickedButton() && 'save_publish' === $form->getClickedButton()->getName()) {
+                        $zenodoClient->publishDeposit($idDeposit,$this->getParameter('app.SBX_TOKEN'));
+                        $action = 'publish';
+                    }
+                    //log user action
+                    $getDepositInfo = json_decode(file_get_contents('https://sandbox.zenodo.org/api/deposit/depositions/'.$idDeposit.'?access_token='.$this->getParameter('app.SBX_TOKEN')),true);
+                    $logInfo = array(
+                        'username' => $userInfo['username'],
+                        'doi_deposit_fix' => $getDepositInfo['conceptrecid'],
+                        'doi_deposit_version' => $getDepositInfo['id'],
+                        'date'=> new \DateTime(),
+                        'action' => $action
+                    );
+                    //addlog return true or exception
+                    $flashMessage = ($action === 'publish') ? $this->addFlash('success', "Successfully completed check here all info : https://sandbox.zenodo.org/record/".$idDeposit) : $this->addFlash('success', "Successfully completed check here all info : https://sandbox.zenodo.org/deposit/".$idDeposit);
+                    $logRepo->addLog($logInfo) ?  $flashMessage :  $this->addFlash('error', 'Something wrong happened');
+                }else{
+                    $error = $postMetadatas->getBody()->getContents();
+                    $logger->debug($error, [
+                        // include extra "context" info in your logs
+                        'context' => 'APICall',
+                    ]);
+                    //api can return two type of array
+                    if (array_key_exists('errors', json_decode($error, true))) {
+                        foreach (json_decode($error, true)['errors'] as $value) {
+                            $this->addFlash('error', $value["message"]);
+                        }
+                    } else {
+                        $this->addFlash('error', json_decode($error, true)['message']);
+                    }
                 }
-                //log user action
-                $getDepositInfo = json_decode(file_get_contents('https://sandbox.zenodo.org/api/deposit/depositions/'.$idDeposit.'?access_token='.$this->getParameter('app.SBX_TOKEN')),true);
-                $logInfo = array(
-                    'username' => $userInfo['username'],
-                    'doi_deposit_fix' => $getDepositInfo['conceptrecid'],
-                    'doi_deposit_version' => $getDepositInfo['id'],
-                    'date'=> new \DateTime(),
-                    'action' => $action
-                );
-                //addlog return true or exception
-                $flashMessage = ($action === 'publish') ? $this->addFlash('success', "Successfully completed check here all info : https://sandbox.zenodo.org/record/".$idDeposit) : $this->addFlash('success', "Successfully completed check here all info : https://sandbox.zenodo.org/deposit/".$idDeposit);
-                $logRepo->addLog($logInfo) ?  $flashMessage :  $this->addFlash('error', 'Something wrong happened');
             }
         }
         return $this->renderForm('deposit/index.html.twig', [
@@ -80,90 +93,103 @@ class DepositController extends AbstractController
         ]);
     }
 
-    public function edit(Request $request, $id, Security $security, ZenodoClient $zenodoClient, UploadFile $uploadFile) : Response {
-
+    public function edit(Request $request, $id, Security $security, ZenodoClient $zenodoClient, UploadFile $uploadFile, LoggerInterface $logger) : Response {
         $userInfo = $security->getToken()->getAttributes();
-
         $token = $this->getParameter('app.SBX_TOKEN');
-
         $response = $zenodoClient->getDepositById($id,$token);
-
-        $depositInfo = json_decode($response->getBody()->getContents(),true);
-
-        $originalId = json_decode($response->getBody(), true)['record_id'];
-
-
-        $fileInfo = $zenodoClient->formatFilesInfoFromDeposit($depositInfo['files']);
-
-        $reformatDepositInfo = $zenodoClient->formatMetadatasFromDeposit($depositInfo);
-
-        $statusDeposit = $depositInfo['submitted'];
-
-        $form = $this->createForm(DepositFormType::class, $reformatDepositInfo);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $deposit = $form->getData();
-            $depositFile = $form->get('depositFile')->getData();
-            if ($form->getClickedButton() && 'new_version' === $form->getClickedButton()->getName()) {
-                $newVersion = $zenodoClient->newVersionDeposit($token,$depositInfo['id']);
-                if ($newVersion['status'] === 201 ) {
-                    return $this->redirectToRoute('edit_deposit', ['id' => $newVersion['idNewVersion']]);
-                } else {
-                    $this->addFlash('error', "An error occurred".$newVersion['message'] );
-                    return $this->redirect($request->getUri());
+        if ($response->getStatusCode() === 200){
+            $depositInfo = json_decode($response->getBody()->getContents(),true);
+            $originalId = json_decode($response->getBody(), true)['record_id'];
+            $fileInfo = $zenodoClient->formatFilesInfoFromDeposit($depositInfo['files']);
+            $reformatDepositInfo = $zenodoClient->formatMetadatasFromDeposit($depositInfo);
+            $statusDeposit = $depositInfo['submitted'];
+            $form = $this->createForm(DepositFormType::class, $reformatDepositInfo);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $deposit = $form->getData();
+                $depositFile = $form->get('depositFile')->getData();
+                if ($form->getClickedButton() && 'new_version' === $form->getClickedButton()->getName()) {
+                    $newVersion = $zenodoClient->newVersionDeposit($token,$depositInfo['id']);
+                    if ($newVersion['status'] === 201 ) {
+                        return $this->redirectToRoute('edit_deposit', ['id' => $newVersion['idNewVersion']]);
+                    } else {
+                        $this->addFlash('error', "An error occurred".$newVersion['message'] );
+                        return $this->redirect($request->getUri());
+                    }
                 }
-            }
-            if ($depositFile) {
-                // foreach pour prévenir des multiple upload
-                $uploadFile->uploadFileLocally($this->getParameter('deposit_upload_directory'), $depositFile);
-                foreach ($depositFile as $fileInfo) {
-                    $filename = $fileInfo->getClientOriginalName();
-                    $path = $this->getParameter('deposit_upload_directory');
-                    $handle = fopen($path.'/'.$filename,'rb');
-                    $bucket = json_decode($response->getBody(), true)['links']['bucket'];
-                    $requestFile =  new guzzleClient();
-                    $requestFile = $requestFile->request('PUT',"$bucket/$filename",[
-                        'query'=> [
-                            'access_token'=>$this->getParameter('app.SBX_TOKEN')
-                        ],
-                        'body' =>  $handle,
-                    ]);
+                if ($depositFile) {
+                    // foreach pour prévenir des multiple upload
+                    $uploadFile->uploadFileLocally($this->getParameter('deposit_upload_directory'), $depositFile);
+                    foreach ($depositFile as $fileInfo) {
+                        $filename = $fileInfo->getClientOriginalName();
+                        $path = $this->getParameter('deposit_upload_directory');
+                        $handle = fopen($path.'/'.$filename,'rb');
+                        $bucket = json_decode($response->getBody(), true)['links']['bucket'];
+                        $requestFile =  new guzzleClient();
+                        $requestFile = $requestFile->request('PUT',"$bucket/$filename",[
+                            'query'=> [
+                                'access_token'=>$this->getParameter('app.SBX_TOKEN')
+                            ],
+                            'body' =>  $handle,
+                        ]);
+                    }
                 }
+                // ajout des données après l'ajout du fichier (si il y en a un)
+                $zenodoClient->postMetadataInDeposit($deposit,$originalId,$token);
+                $idDeposit = json_decode($response->getBody(),true)['id'];
+                // Define the action for the log
+                $action = 'save';
+                if ($form->getClickedButton() && 'save_publish' === $form->getClickedButton()->getName()) {
+                    $publishDeposit = $zenodoClient->publishDeposit($idDeposit,$this->getParameter('app.SBX_TOKEN'));
+                    if ($publishDeposit['status'] !== 202) {
+                        $this->addFlash('error', $publishDeposit['message']);
+                        return $this->redirect($request->getUri());
+                    }
+                    $action = 'publish';
+                }
+                $getDepositInfo = json_decode(file_get_contents('https://sandbox.zenodo.org/api/deposit/depositions/'.$idDeposit.'?access_token='.$this->getParameter('app.SBX_TOKEN')),true);
+                $logInfo = array(
+                    'username' => $userInfo['username'],
+                    'doi_deposit_fix' => $getDepositInfo['conceptrecid'],
+                    'doi_deposit_version' => $getDepositInfo['id'],
+                    'date'=> new \DateTime(),
+                    'action' => $action,
+                );
+                $statusDeposit = $getDepositInfo['submitted'];
+
+                $flashMessage = ($action === 'publish') ? $this->addFlash('success', "Successfully completed check here all info : https://sandbox.zenodo.org/record/".$idDeposit) : $this->addFlash('success', "Successfully completed check here all info : https://sandbox.zenodo.org/deposit/".$idDeposit);
+                return $this->redirect($request->getUri());
+            } else {
+                return $this->renderForm('deposit/edit.html.twig', [
+                    'controller_name' => 'DepositController',
+                    'form' => $form,
+                    'filesInfo' => $fileInfo,
+                    'idDeposit' => $id,
+                    'DepositPublished' => $statusDeposit,
+                    'userInfo' => [
+                        'lastname' => $userInfo['LASTNAME'],
+                        'firstname' => $userInfo['FIRSTNAME'],
+                    ]
+                ]);
             }
-
-            $pushMeta =  new guzzleClient();
-
-            // ajout des données après l'ajout du fichier (si il y en a un)
-
-            $zenodoClient->postMetadataInDeposit($deposit,$originalId,$token);
-
-            $idDeposit = json_decode($response->getBody(),true)['id'];
-            // Define the action for the log
-            $action = 'save';
-            if ($form->getClickedButton() && 'save_publish' === $form->getClickedButton()->getName()) {
-                $zenodoClient->publishDeposit($idDeposit,$this->getParameter('app.SBX_TOKEN'));
-                $action = 'publish';
-            }
-            $getDepositInfo = json_decode(file_get_contents('https://sandbox.zenodo.org/api/deposit/depositions/'.$idDeposit.'?access_token='.$this->getParameter('app.SBX_TOKEN')),true);
-            $logInfo = array(
-                'username' => $userInfo['username'],
-                'doi_deposit_fix' => $getDepositInfo['conceptrecid'],
-                'doi_deposit_version' => $getDepositInfo['id'],
-                'date'=> new \DateTime(),
-                'action' => $action,
-            );
-            $statusDeposit = $getDepositInfo['submitted'];
-
-            $flashMessage = ($action === 'publish') ? $this->addFlash('success', "Successfully completed check here all info : https://sandbox.zenodo.org/record/".$idDeposit) : $this->addFlash('success', "Successfully completed check here all info : https://sandbox.zenodo.org/deposit/".$idDeposit);
-            return $this->redirect($request->getUri());
         } else {
+            $exceptionMessage = '';
+            switch ($response->getStatusCode()){
+                case 404:
+                    $exceptionMessage = 'Deposit not found be sure you have the right informations';
+                    break;
+                default:
+                    $exceptionMessage = json_decode($response->getBody()->getContents(),true)['message'];
+                    break;
+            }
+            $logger->debug($response->getBody()->getContents(), [
+                // include extra "context" info in your logs
+                'context' => 'APICall',
+            ]);
 
-            return $this->renderForm('deposit/edit.html.twig', [
-                'controller_name' => 'DepositController',
-                'form' => $form,
-                'filesInfo' => $fileInfo,
-                'idDeposit' => $id,
-                'DepositPublished' => $statusDeposit,
+            return $this->render('zenodoexception/error.html.twig',[
+                'statusCode' => $response->getStatusCode(),
+                'message' => $exceptionMessage,
                 'userInfo' => [
                     'lastname' => $userInfo['LASTNAME'],
                     'firstname' => $userInfo['FIRSTNAME'],
@@ -173,30 +199,43 @@ class DepositController extends AbstractController
     }
 
     public function deleteFile (Request $request, Security $security, ZenodoClient $zenodoClient, $id, $fileId, Session $session) {
-
         if ($security->getToken()->getAttributes()){
             $token = $this->getParameter('app.SBX_TOKEN');
             $fileInfoSended = json_decode($request->getContent(), true);
             $deposit = $zenodoClient->getDepositById($id,$token);
-            $depositInfo = json_decode($deposit->getBody()->getContents(),true);
-            $checkValidFile = false;
-
-            foreach ($depositInfo['files'] as $key => $value){
-                if (in_array($fileInfoSended['checksum'],$depositInfo['files'][$key],true)
-                    && (in_array($fileInfoSended['linkFile'],$depositInfo['files'][$key]['links'],true))
-                    && (in_array($fileInfoSended['fileName'],$depositInfo['files'][$key],true))
-                    && (in_array($fileInfoSended['id'],$depositInfo['files'][$key],true)))  {
+            if ($deposit->getStatusCode() === 200) {
+                $depositInfo = json_decode($deposit->getBody()->getContents(),true);
+                $checkValidFile = false;
+                foreach ($depositInfo['files'] as $key => $value){
+                    if (in_array($fileInfoSended['checksum'],$depositInfo['files'][$key],true)
+                        && (in_array($fileInfoSended['linkFile'],$depositInfo['files'][$key]['links'],true))
+                        && (in_array($fileInfoSended['fileName'],$depositInfo['files'][$key],true))
+                        && (in_array($fileInfoSended['id'],$depositInfo['files'][$key],true)))  {
                         $checkValidFile = true;
                         break;
                     }
+                }
+                if ($checkValidFile === true){
+                    $deleteFile = $zenodoClient->deleteFilesFromDeposit($token,$id,$fileId);
+                }else{
+                    return new JsonResponse([
+                        'status' => 404,
+                        'message' => 'File not found'
+                    ]);
+                }
+                return new JsonResponse($deleteFile);
+            } else {
+                $error = $deposit->getBody()->getContents();
+                if (array_key_exists('errors', json_decode($error, true))) {
+                    foreach (json_decode($error, true)['errors'] as $value) {
+                        $this->addFlash('error', $value["message"]);
+                    }
+                } else {
+                    return new JsonResponse('error', json_decode($error, true)['message']);
+                }
             }
-            if ($checkValidFile === true){
-                $deleteFile = $zenodoClient->deleteFilesFromDeposit($token,$id,$fileId);
-            }
-            return new JsonResponse($deleteFile);
-        }
-        else{
-            return new JsonResponse('no results found', Response::HTTP_FORBIDDEN);
+        } else {
+            return new JsonResponse('You must be authentified to make this action', Response::HTTP_FORBIDDEN);
         }
     }
 }
